+++ core-project/asakusa-runtime-all/pom.xml
@@ -66,6 +66,11 @@
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
+      <artifactId>asakusa-runtime-tsv</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
       <artifactId>asakusa-runtime-inprocess</artifactId>
       <version>${project.version}</version>
     </dependency>
