+++ core-project/tsv/.gitignore
@@ -1,0 +1,5 @@
+/.settings
+/target
+/.classpath
+/.project
+/bin
