+++ core-project/tsv/src/main/java/com/asakusafw/runtime/io/tsv/DelimiterRangeInputStream.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.asakusafw.runtime.directio.util;
+package com.asakusafw.runtime.io.tsv;
 
 import java.io.IOException;
 import java.io.InputStream;
