+++ core-project/tsv/src/main/java/com/asakusafw/runtime/io/tsv/TsvConstants.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.asakusafw.runtime.io;
+package com.asakusafw.runtime.io.tsv;
 
 /**
  * The constants about TSV format.
