+++ core-project/tsv/src/test/java/com/asakusafw/runtime/io/tsv/DelimiterRangeInputStreamTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.asakusafw.runtime.directio.util;
+package com.asakusafw.runtime.io.tsv;
 
 import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.*;
