+++ core-project/tsv/src/test/resources/com/asakusafw/runtime/io/tsv/int
