+++ sandbox-project/asakusa-windgate-dmdl-ext/src/main/java/com/asakusafw/dmdl/windgate/stream/driver/StreamSupportEmitter.java
@@ -40,8 +40,8 @@ import com.asakusafw.dmdl.semantics.type.BasicType;
 import com.asakusafw.dmdl.windgate.util.FsProcessDescriptionGenerator;
 import com.asakusafw.runtime.io.RecordEmitter;
 import com.asakusafw.runtime.io.RecordParser;
-import com.asakusafw.runtime.io.TsvEmitter;
-import com.asakusafw.runtime.io.TsvParser;
+import com.asakusafw.runtime.io.tsv.TsvEmitter;
+import com.asakusafw.runtime.io.tsv.TsvParser;
 import com.asakusafw.utils.java.model.syntax.ClassDeclaration;
 import com.asakusafw.utils.java.model.syntax.ExpressionStatement;
 import com.asakusafw.utils.java.model.syntax.FieldDeclaration;
